+++ client/src/components/BuyProduct/Payment.jsx
@@ -1,10 +1,29 @@
 import React from 'react';
+import Invoice from './Invoice';
+import { ToastContainer, toast } from 'react-toastify';
 
 const Payment = ({
     name, email,
     countryCode, phoneNo,
-    address, total
+    address, total,
+    products, method
 }) => {
+
+    const cashOnDelivery = () => {
+        Invoice({
+            name,
+            countryCode,
+            phoneNo,
+            email,
+            address,
+            products,
+            total,
+            method: 'cod',
+            transactionId: null
+        });
+        toast.success("Order placed Successfully!!", { autoClose: 3000 });
+    };
+
 
     const loadRazorpay = () => {
         const script = document.createElement('script');
@@ -25,10 +44,22 @@ const Payment = ({
                 currency: "INR",
                 name: 'Enactus NIT Prayagraj',
                 description: 'Test Transaction',
+
                 handler: function (response) {
-                    console.log(response);
-                    alert("Payment successful!!");
+                    Invoice({
+                        name,
+                        email,
+                        address,
+                        phoneNo,
+                        countryCode,
+                        products,
+                        total,
+                        method: 'online',
+                        transactionId: response.razorpay_payment_id
+                    });
+                    toast.success("Payment Compleleted Successful!!", { autoClose: 3000 });
                 },
+
                 prefill: {
                     name: name,
                     email: email,
@@ -55,13 +86,24 @@ const Payment = ({
 
     return (
 
-        <div className="text-center pt-4">
-            <button
-                onClick={loadRazorpay}
-                className="bg-yellow-600 hover:bg-yellow-700 text-white px-4 py-2 rounded"
-            >
-                Pay with Razorpay
-            </button>
+        <div className="mt-6 text-center">
+            {method === 'online' ? (
+                <button
+                    onClick={loadRazorpay}
+                    className="bg-yellow-600 hover:bg-yellow-700 text-white px-6 py-2 rounded"
+                >
+                    Pay Online (via Razorpay)
+                </button>
+            ) : (
+                <button
+                    onClick={cashOnDelivery}
+                    className="bg-green-600 hover:bg-green-700 text-white px-6 py-2 rounded"
+                >
+                    Place Order (Cash on Delivery)
+                </button>
+            )}
+            
+            <ToastContainer />
         </div>
     );
 };
